UpdateResource logic is fixed
