+++ src/main/java/rueppellii/backend2/tribes/resource/service/ResourceService.java
@@ -2,6 +2,7 @@ package rueppellii.backend2.tribes.resource.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import rueppellii.backend2.tribes.building.persistence.model.Building;
 import rueppellii.backend2.tribes.building.service.BuildingService;
 import rueppellii.backend2.tribes.gameUtility.timeService.TimeService;
 import rueppellii.backend2.tribes.kingdom.persistence.model.Kingdom;
@@ -16,21 +17,22 @@ import rueppellii.backend2.tribes.resource.exception.NoResourceException;
 import java.util.ArrayList;
 import java.util.List;
 
+import static java.util.Objects.requireNonNull;
 import static rueppellii.backend2.tribes.gameUtility.timeService.TimeConstants.ONE_MINUTE_IN_SECONDS;
+import static rueppellii.backend2.tribes.resource.utility.ResourceConstants.RESOURCE_PER_MINUTE_BUILDING_LEVEL_MULTIPLIER;
+import static rueppellii.backend2.tribes.resource.utility.ResourceConstants.RESOURCE_PER_MINUTE_BUILDING_LEVEL_ONE;
 import static rueppellii.backend2.tribes.resource.utility.ResourceFactory.makeResource;
 
 @Service
 public class ResourceService {
     private ResourceRepository resourceRepository;
     private TimeService timeService;
-    private BuildingService buildingService;
     private KingdomService kingdomService;
 
     @Autowired
-    public ResourceService(ResourceRepository resourceRepository, TimeService timeService, BuildingService buildingService, KingdomService kingdomService) {
+    public ResourceService(ResourceRepository resourceRepository, TimeService timeService, KingdomService kingdomService) {
         this.resourceRepository = resourceRepository;
         this.timeService = timeService;
-        this.buildingService = buildingService;
         this.kingdomService = kingdomService;
     }
 
@@ -61,7 +63,7 @@ public class ResourceService {
         List<Resource> resources = kingdom.getKingdomsResources();
         for (Resource r : resources) {
             Integer baseResourcePerMinute = r.getResourcePerMinute();
-            Integer totalBuildingLevelMultiplier = buildingService.getTotalBuildingLevelMultiplier(kingdom.getKingdomsBuildings(), r);
+            Integer totalBuildingLevelMultiplier = getTotalBuildingLevelMultiplier(kingdom.getKingdomsBuildings(), r);
             Integer elapsedSeconds = timeService.calculateElapsedSeconds(r.getUpdatedAt());
             Long remainderSeconds = timeService.calculateRemainder(r.getUpdatedAt());
             if (r instanceof Gold && calculateGold(baseResourcePerMinute, totalBuildingLevelMultiplier, elapsedSeconds) != 0) {
@@ -74,6 +76,25 @@ public class ResourceService {
             }
         }
         kingdomService.save(kingdom);
+    }
+
+    public Integer getTotalBuildingLevelMultiplier(List<Building> kingdomsBuildings, Resource resource) {
+        String buildingName = null;
+        if (resource instanceof Gold) {
+            buildingName = "MINE";
+        }
+        if (resource instanceof Food) {
+            buildingName = "FARM";
+        }
+        String finalBuildingName = buildingName;
+        Integer numberOfBuildings = (int) kingdomsBuildings.stream().filter(building -> building.getType().getName().matches(requireNonNull(finalBuildingName))).count();
+        Integer levelOneBuildingMultiplier = numberOfBuildings * RESOURCE_PER_MINUTE_BUILDING_LEVEL_ONE;
+        Integer totalLevelOfBuildings = kingdomsBuildings.stream().filter(building -> building.getType().getName().matches(finalBuildingName)).mapToInt(Building::getLevel).sum();
+        if (!numberOfBuildings.equals(totalLevelOfBuildings)) {
+            Integer totalBuildingLevelMultiplier = RESOURCE_PER_MINUTE_BUILDING_LEVEL_MULTIPLIER * (totalLevelOfBuildings - numberOfBuildings);
+            return levelOneBuildingMultiplier + totalBuildingLevelMultiplier;
+        }
+        return levelOneBuildingMultiplier;
     }
 
     private Integer calculateFood(Kingdom kingdom, Integer baseResourcePerMinute,
